+++ doc/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 Version 0.9.2
 -------------
 
+* Add :meth:`Runner.timeit` method.
 * Fix ``stats`` command: display again statistics on the whole benchmark suite.
 * Fix a ResourceWarning if interrupted:  Runner now kills the worker process
   when interrupted.